launch file for offline phase
 This launch file takes care of launching the mapping, rssi-pose and alignment
components. For the offline phase you'll need two invocations. The first one to
create the .pgm map (only hand over the bagfile argument):

$ roslaunch rssimap offline-phase.launch bagfile:=$(pwd)/rssimap_2012-11-28-16-34-01

Notice that the argument needs to given as an absolute path and without the .bag
file ending!

The second step cares for aligning multiple maps and extracting the .csv file
which contains localized rssi samples. For this you need a "world" map, i.e. a
map to align others too and then start offline phase with:

 $ roslaunch rssimap offline_phase.launch bagfile:=$(pwd)/rssimap_2012-11-28-16-34-01 doalign:=1 worldmap:=$(pwd)/world.yaml

This creates a .csv file next to the bagfile.